Refactor main/commit() into a separate file
